+++ pom.xml
@@ -1,11 +1,16 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>com.extendedclip.papiexpansion.example</groupId>
-  <artifactId>example-expansion</artifactId>
-  <version>0.0.1-SNAPSHOT</version>
-  <name>PAPI-Expansion-Example</name>
-  <description>This is an example expansion for PlaceholderAPI</description>
-    
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+
+    <modelVersion>4.0.0</modelVersion>
+
+    <groupId>com.extendedclip.papiexpansion.example</groupId>
+    <artifactId>example-expansion</artifactId>
+    <version>0.0.1-SNAPSHOT</version>
+    <packaging>jar</packaging>
+
+    <name>PAPI-Expansion-Example</name>
+    <description>This is an example expansion for PlaceholderAPI</description>
+
     <repositories>
         <repository>
             <id>spigot-repo</id>
@@ -20,19 +25,19 @@
             <url>http://repo.extendedclip.com/content/repositories/public/</url>
         </repository>
     </repositories>
-    
+
     <dependencies>
         <dependency>
             <groupId>org.spigotmc</groupId>
             <artifactId>spigot-api</artifactId>
-            <version>1.12-R0.1-SNAPSHOT</version>
+            <version>1.12.2-R0.1-SNAPSHOT</version>
             <scope>compile</scope>
         </dependency>
         <dependency>
-         <groupId>me.clip</groupId>
-          <artifactId>placeholderapi</artifactId>
-          <version>2.8.3</version>
-         <scope>provided</scope>
+            <groupId>me.clip</groupId>
+            <artifactId>placeholderapi</artifactId>
+            <version>2.9.1</version>
+            <scope>provided</scope>
         </dependency>
         <dependency>
             <groupId>me.clip.deluxetags</groupId>
@@ -41,8 +46,8 @@
             <scope>provided</scope>
         </dependency>
     </dependencies>
-    
-    <build>     
+
+    <build>
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
@@ -65,5 +70,4 @@
             </plugin>
         </plugins>
     </build>
-    
 </project>
